+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased Changes
 * Added `NonMaxI128` and `NonMaxU128` support
 * Implemented `std::convert::[Try]From` for `NonMax*` & primitive types to match `NonZero*`
+* Implemented `std::fmt::{Display, Binary, Octal, LowerHex, UpperHex}` to match `NonZero*`
+* Removed outer "NonMax(...)" in output from `impl Debug for NonMax*`, matching `NonZero*` behavior
 
 ## 0.4.0 (2020-09-27)
 * Raised MSRV to 1.46.0 to make more methods `const`.